requirements.txt
# Flask and Extensions
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
flask-restx==1.3.0

# Database
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pg8000>=1.31.1

# Google Cloud
cloud-sql-python-connector[pg8000]==1.12.1
google-cloud-secret-manager==2.18.0
google-cloud-tasks==2.14.1
google-cloud-logging==3.9.0

# Maps
googlemaps==4.6.0

# Email
requests==2.31.0  # Required for Mailgun

# JWT & Security
Flask-JWT-Extended>=4.3.0
bcrypt>=3.2.0
cryptography==42.0.5
python-jose==3.3.0

# Utils
uuid>=1.30

# Production Server
gunicorn==21.2.0

# Rate Limiting
Flask-Limiter==3.3.0

# Testing dependencies
pytest==8.0.2
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-flask==1.3.0
coverage==7.4.3

# Linting and Formatting
flake8==3.9.2
black==24.2.0
isort==5.13.2
mypy==1.8.0
autopep8==1.5.7 